Add test_logging.py
